#36 Implement own property collector based on fields (wip)
The current implementation is subject to change. The JavaBeans standard infers the available properties on a bean based on the available methods. It may make more sense to continue using Java's bean introspector to gather the properties, and in a subsequent step, to match the properties with a field where possible and to convert Java's PropertyDescriptor into our new BeanPropertyDescription class with the additional information.